requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
binaryornot==0.4.4
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cssselect2==0.7.0
fastapi==0.115.0
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
fonttools==4.54.1
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pefile==2024.8.26
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyphen==0.16.0
pypng==0.20220715.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.8.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
text-unidecode==1.3
tinycss2==1.3.0
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
weasyprint==62.3
webencodings==0.5.1
websockets==13.1
zopfli==0.2.3
